requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.3
distro==1.9.0
fastapi==0.136.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.12
jiter==0.14.0
openai==2.32.0
pydantic==2.13.3
pydantic_core==2.46.3
PyPDF2==3.0.1
python-dotenv==1.2.2
sniffio==1.3.1
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
